+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### First Release
 
 #### Module ryan_joiner
+- Added ``_normal_order_statistic`` function (05/11/2023);
 - Added ``_critical_value`` function (06/11/2023);
 - Added ``p_value`` function (05/11/2023);
 - Added ``statistic`` function (05/11/2023);