+++ docs/changelog.md
@@ -5,11 +5,11 @@
 
 
 #### Module ryan_joiner
+- Added ``_normal_order_statistic`` function (06/11/2023);
 - Added ``_critical_value`` function (06/11/2023);
 - Added ``p_value`` function (05/11/2023);
 - Added ``statistic`` function (05/11/2023);
 - Added ``citation`` function (04/11/2023);
-- Added ``normal_order_statistic`` function (02/11/2023);
 - Added ``order_statistic`` function (02/11/2023);